refactor: i18n updates to frontend base
