+++ src/account-settings/BetaLanguageBanner.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { Button, Hyperlink } from '@openedx/paragon';
 
 import { betaLanguageBannerSelector } from './data/selectors';
-import messages from './AccountSettingsPage.messages';
+import messages from './messages';
 import { saveSettings } from './data/actions';
 import { TRANSIFEX_LANGUAGE_BASE_URL } from './data/constants';
 import Alert from './Alert';
