+++ src/account-settings/EmailField.jsx
@@ -10,7 +10,7 @@ import { faExclamationTriangle, faPencilAlt } from '@fortawesome/free-solid-svg-
 
 import Alert from './Alert';
 import SwitchContent from './SwitchContent';
-import messages from './AccountSettingsPage.messages';
+import messages from './messages';
 
 import {
   openForm,
